Refactor ErrorBuilders and ErrorCreators tests
*Fixed EqualsBuilder in ValidationError
*Renamed error folder to errorBuilder
*rename error_creators to ErrorCreator
*Modify BrokenRelationErrorCreator description string to be more clear
*Delete tests for the BrokenRelationErrorBuilder to test for missing column and tables. This behavior is redudant.
