+++ src/main/java/org/pdxfinder/validator/tablevalidation/errorBuilders/DuplicateValueErrorBuilder.java
@@ -1,24 +1,24 @@
-package org.pdxfinder.validator.tablevalidation.error;
+package org.pdxfinder.validator.tablevalidation.errorBuilders;
 
 import org.pdxfinder.validator.tablevalidation.dto.ValidationError;
 import org.pdxfinder.validator.tablevalidation.enums.ErrorType;
 
-public class MissingColumnErrorBuilder extends ValidationErrorBuilder {
+public class DuplicateValueErrorBuilder extends ValidationErrorBuilder {
 
-  private String errorType = ErrorType.MISSING_COLUMN.getErrorType();
-  private String tablename;
+  private String errorType = ErrorType.DUPLICATE_VALUES.getErrorType();
+  private String tableName;
   private String columnName;
-  private String cause = "Required column not found in table";
+  private String cause = "";
   private String rule = "";
 
-  public MissingColumnErrorBuilder(String tableName, String columnName) {
-    this.tablename = tableName;
+  public DuplicateValueErrorBuilder(String tableName, String columnName) {
+    this.tableName = tableName;
     this.columnName = columnName;
   }
 
   @Override
-  public ValidationErrorBuilder<?> buildCause(String cause) {
-    this.cause = cause;
+  public ValidationErrorBuilder<?> buildCause(String duplicates) {
+    this.cause = String.format("Duplicated values found: %s", duplicates);
     return this;
   }
 
@@ -28,13 +28,13 @@ public class MissingColumnErrorBuilder extends ValidationErrorBuilder {
     return this;
   }
 
-  @Override
   public ValidationError build() {
     return new ValidationError.Builder(errorType)
-            .setTableName(tablename)
+            .setTableName(tableName)
             .setColumnName(columnName)
             .setCause(cause)
             .setRule(rule)
             .build();
+
   }
 }
