+++ src/main/java/org/pdxfinder/validator/tablevalidation/errorBuilders/MissingValueErrorBuilder.java
@@ -1,17 +1,20 @@
-package org.pdxfinder.validator.tablevalidation.error;
+package org.pdxfinder.validator.tablevalidation.errorBuilders;
 
 import org.pdxfinder.validator.tablevalidation.dto.ValidationError;
 import org.pdxfinder.validator.tablevalidation.enums.ErrorType;
 
-public class MissingTableErrorBuilder extends ValidationErrorBuilder {
+public class MissingValueErrorBuilder extends ValidationErrorBuilder {
 
-    private String errorType = ErrorType.MISSING_COLUMN.getErrorType();
+    private String errorType = ErrorType.MISSING_VALUE.getErrorType();
     private String tableName;
-    private String cause = "Required table is not found";
-    private String rule;
+    private String columnName;
+    private String rule = "Required value";
+    private String cause = "No value found";
+    private String rows;
 
-    public MissingTableErrorBuilder(String tableName) {
+    public MissingValueErrorBuilder(String tableName, String columnName) {
         this.tableName = tableName;
+        this.columnName = columnName;
     }
 
     @Override
@@ -26,12 +29,19 @@ public class MissingTableErrorBuilder extends ValidationErrorBuilder {
         return this;
     }
 
+    public MissingValueErrorBuilder buildRows(String rows) {
+        this.rows = rows;
+        return this;
+    }
+
     @Override
     public ValidationError build() {
         return new ValidationError.Builder(errorType)
                 .setTableName(tableName)
+                .setColumnName(columnName)
                 .setRule(rule)
                 .setCause(cause)
+                .setRow(rows)
                 .build();
     }
 }
