+++ src/test/java/org/pdxfinder/validator/tablevalidation/errorBuilders/DuplicateValueErrorBuilderTest.java
@@ -1,9 +1,9 @@
-package org.pdxfinder.validator.tablevalidation.error;
+package org.pdxfinder.validator.tablevalidation.errorBuilders;
 
 import org.junit.Test;
 import org.pdxfinder.validator.tablevalidation.dao.ColumnReference;
 import org.pdxfinder.validator.tablevalidation.dto.ValidationError;
-import org.pdxfinder.validator.tablevalidation.error_creators.DuplicateValueErrorCreator;
+import org.pdxfinder.validator.tablevalidation.errorCreators.DuplicateValueErrorCreator;
 
 import java.util.Arrays;
 import java.util.HashSet;
@@ -17,7 +17,7 @@ public class DuplicateValueErrorBuilderTest {
 
   @Test
   public void message() {
-    String expected = "Duplicates values found: [a]";
+    String expected = "Duplicated values found: [a]";
     ColumnReference uniqueColumn = ColumnReference.of("table", "column");
     Set<String> duplicateValues = new HashSet<>(Arrays.asList("a"));
     ValidationError error =
