+++ src/test/java/org/pdxfinder/validator/tablevalidation/errorBuilders/MissingColumnErrorBuilderTest.java
@@ -1,10 +1,10 @@
-package org.pdxfinder.validator.tablevalidation.error;
+package org.pdxfinder.validator.tablevalidation.errorBuilders;
 
 import org.junit.Test;
 import org.pdxfinder.validator.tablevalidation.dao.ColumnReference;
 import org.pdxfinder.validator.tablevalidation.dto.ValidationError;
 import org.pdxfinder.validator.tablevalidation.enums.ErrorType;
-import org.pdxfinder.validator.tablevalidation.error_creators.MissingColumnErrorCreator;
+import org.pdxfinder.validator.tablevalidation.errorCreators.MissingColumnErrorCreator;
 
 import static org.junit.Assert.assertEquals;
 
@@ -14,7 +14,7 @@ public class MissingColumnErrorBuilderTest {
 
   @Test
   public void message_givenMissingValue_returnsAppropriateError() {
-    String expected = "Missing column: [column]";
+    String expected = "Required column not found in table";
     String expected_table = "table1";
     ColumnReference columnReference = ColumnReference.of(expected_table, "column");
     ValidationError error = missingColumnErrorCreator.create(columnReference);
